refactor: move ReactionService up a level and subclass for fulfillment
