+++ imports/services/fulfillment/core/index.js
@@ -1,0 +1,6 @@
+import graphqlSchema from "./schema.graphql";
+
+export const graphqlSchemas = [graphqlSchema];
+export { default as graphqlResolvers } from "./resolvers";
+export { default as ReactionFulfillmentService } from "./ReactionFulfillmentService";
+export { default as startup } from "./startup";
